% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QUALYPSO.r
\name{checkTypeandDimension}
\alias{checkTypeandDimension}
\title{checkTypeandDimension}
\usage{
checkTypeandDimension(Y, scen, X)
}
\arguments{
\item{Y}{matrix \code{nS} x \code{nY}}

\item{scen}{data.frame \code{nS} x \code{nEff} with the \code{nEff} characteristics
(e.g. type of GCM) for each of the \code{nS} scenarios. The number of characteristics
 \code{nEff} corresponds to the number of main effects that will be included in the ANOVA model.}

\item{X}{(optional) predictors corresponding to the projections, e.g. time or global temperature.
It can be a vector if the predictor is the same for all scenarios (e.g. \code{X=2001:2100}) or
a matrix of the same size as Y if these predictors are different for the scenarios. By default,
a vector \code{1:nY} is created.}
}
\description{
Check type and dimensions if objects passed as arguments
}
\author{
Guillaume Evin
}
