% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataQUALYPSO-data.r
\docType{data}
\name{dataQUALYPSO}
\alias{dataQUALYPSO}
\title{Example of climate projections}
\format{matrix of climate projections: 3 scenarios x 100 years}
\usage{
data(dataQUALYPSO)
}
\description{
Example of climate projections for two fictive RCMs and two fictive GCMs, for 100 years.
}
\references{
Evin, G., B. Hingray, J. Blanchet, N. Eckert, S. Morin, and D. Verfaillie. (2019) <doi:10.1175/JCLI-D-18-0606.1>.
}
\author{
Guillaume Evin \email{guillaume.evin@irstea.fr}
}
\keyword{datasets}
