% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QUALYPSO.r
\name{get.Qmat}
\alias{get.Qmat}
\title{get.Qmat}
\usage{
get.Qmat(p)
}
\arguments{
\item{p}{integer}
}
\value{
\item{matrix}{p x p matrix}
}
\description{
Provide matrix Q derived from a matrix of Helmert contrasts: \deqn{ Q = Q^* (Q^{*T} Q^*)^{-1/2}}
See Eq. A6 in Evin et al. (2019)
}
\author{
Guillaume Evin
}
