% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QUALYPSO.r
\name{get.Qstar.mat}
\alias{get.Qstar.mat}
\title{get.Qstar.mat}
\usage{
get.Qstar.mat(p)
}
\arguments{
\item{p}{integer}
}
\value{
\item{matrix}{p x (p-1) matrix containing Helmert contrasts}
}
\description{
Provide matrix containing Helmert contrasts (see Eq. A10 in Tingley et al., 2011).
}
\references{
Tingley, Martin P. 2011. A Bayesian ANOVA Scheme for Calculating Climate Anomalies,
with Applications to the Instrumental Temperature Record. Journal of Climate 25 (2). https://doi.org/10.1175/JCLI-D-11-00008.1.
}
\author{
Guillaume Evin
}
