% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QUALYPSO.r
\name{plotQUALYPSOclimateChangeResponse}
\alias{plotQUALYPSOclimateChangeResponse}
\title{plotQUALYPSOclimateChangeResponse}
\usage{
plotQUALYPSOclimateChangeResponse(
  QUALYPSOOUT,
  lim = NULL,
  xlab = "",
  ylab = "Climate change response",
  ...
)
}
\arguments{
\item{QUALYPSOOUT}{output from \code{\link{QUALYPSO}}}

\item{lim}{y-axis limits (default is NULL)}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{...}{additional arguments to be passed to \code{\link[graphics]{plot}}}
}
\description{
Plot climate change responses.
}
\author{
Guillaume Evin
}
