% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QUALYPSO.r
\name{plotQUALYPSOclimateResponse}
\alias{plotQUALYPSOclimateResponse}
\title{plotQUALYPSOclimateResponse}
\usage{
plotQUALYPSOclimateResponse(
  QUALYPSOOUT,
  lim = NULL,
  xlab = "X",
  ylab = "Y",
  iS = NULL,
  ...
)
}
\arguments{
\item{QUALYPSOOUT}{output from \code{\link{QUALYPSO}}}

\item{lim}{y-axis limits (default is NULL)}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{iS}{index of the projection (integer). If NULL, one scenario is chosen randomly}

\item{...}{additional arguments to be passed to \code{\link[graphics]{plot}}}
}
\description{
Plot the climate responses.
}
\author{
Guillaume Evin
}
