% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QUALYPSO.r
\name{plotQUALYPSOeffect}
\alias{plotQUALYPSOeffect}
\title{plotQUALYPSOeffect}
\usage{
plotQUALYPSOeffect(QUALYPSOOUT, iEff, includeMean = FALSE, iMedian = 7,
  iBinf = 2, iBsup = 12, lim = NULL, col = 1:20, xlab = "Years",
  ylab = "Effect", addLegend = TRUE, ...)
}
\arguments{
\item{QUALYPSOOUT}{output from \code{\link{QUALYPSO}}}

\item{iEff}{index of the main effect to be plotted in \code{QUALYPSOOUT$listScenarioInput$listEff}}

\item{includeMean}{if TRUE, the grand mean is added to the main effect in the plot}

\item{iMedian}{index corresponding to the median (Pr=0.5) in \code{QUALYPSOOUT$listOption$quantileCompress} (default=7)}

\item{iBinf}{index corresponding to the lower bound in \code{QUALYPSOOUT$listOption$quantileCompress} (default=2)}

\item{iBsup}{index corresponding to the upper bound in \code{QUALYPSOOUT$listOption$quantileCompress} (default=12)}

\item{lim}{y-axis limits (default is NULL)}

\item{col}{colors for each effect}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{addLegend}{if TRUE, a legend is added}

\item{...}{additional arguments to be passed to \code{\link[graphics]{plot}}}
}
\description{
Plot prediction of ANOVA effects for one main effect. By default, we plot we plot the credible intervals corresponding to a probability 0.95.
}
\author{
Guillaume Evin
}
