% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QUALYPSO.r
\name{plotQUALYPSOgrandmean}
\alias{plotQUALYPSOgrandmean}
\title{plotQUALYPSOgrandmean}
\usage{
plotQUALYPSOgrandmean(
  QUALYPSOOUT,
  lim = NULL,
  col = "black",
  xlab = "",
  ylab = "Grand mean",
  addLegend = T,
  ...
)
}
\arguments{
\item{QUALYPSOOUT}{output from \code{\link{QUALYPSO}}}

\item{lim}{y-axis limits (default is NULL)}

\item{col}{color for the overall mean and the credible interval}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{addLegend}{if TRUE, a legend is added}

\item{...}{additional arguments to be passed to \code{\link[graphics]{plot}}}
}
\description{
Plot prediction of grand mean ensemble.
}
\author{
Guillaume Evin
}
