% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QUALYPSO.r
\name{plotQUALYPSOTotalVarianceByScenario}
\alias{plotQUALYPSOTotalVarianceByScenario}
\title{plotQUALYPSOTotalVarianceByScenario}
\usage{
plotQUALYPSOTotalVarianceByScenario(
  QUALYPSOOUT,
  nameEff,
  nameScenario,
  col = NULL,
  ylim = NULL,
  xlab = "",
  ylab = "Change variable",
  addLegend = TRUE,
  ...
)
}
\arguments{
\item{QUALYPSOOUT}{output from \code{\link{QUALYPSO}}}

\item{nameEff}{name of the main effect to be plotted in \code{QUALYPSOOUT$namesEff}}

\item{nameScenario}{name of the scenario to be plotted (as provided in \code{scenAvail})}

\item{col}{colors for each source of uncertainty, the first two colors
corresponding to internal variability and residual variability, respectively}

\item{ylim}{y-axis limits}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{addLegend}{if TRUE, a legend is added}

\item{...}{additional arguments to be passed to \code{\link[graphics]{plot}}}
}
\description{
Plot fraction of total variance explained by each source of uncertainty.
}
\author{
Guillaume Evin
}
