% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QUALYPSO.r
\name{QUALYPSO.check.option}
\alias{QUALYPSO.check.option}
\title{QUALYPSO.check.option}
\usage{
QUALYPSO.check.option(listOption)
}
\arguments{
\item{listOption}{list of options}
}
\value{
listOption
}
\description{
Check possible options
}
\author{
Guillaume Evin
}
