% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QUALYPSO.r
\name{QUALYPSO.process.scenario}
\alias{QUALYPSO.process.scenario}
\title{QUALYPSO.process.scenario}
\usage{
QUALYPSO.process.scenario(scen)
}
\arguments{
\item{scen}{data.frame \code{nS} x \code{nEff} with the \code{nEff} characteristics (e.g. type of GCM) for each of the \code{nS} x \code{nS} scenarios}
}
\value{
list of preprocessed objects (\code{listEff, scen, scenComp, nEff, nTypeEff, nComp, isMissing, nMissing, iMatchScen,
indexEffInCompScen, Qmat})
}
\description{
Process input scenarios.
}
\author{
Guillaume Evin
}
