% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/description-data.r
\docType{data}
\name{scen_DJFTas}
\alias{scen_DJFTas}
\title{data.frame indicating which GCMs and RCMs have been used for the 20 climate projections
of mean winter temperature over CEU}
\format{
data.frame with 20 rows and two columns: GCM and RCM
}
\usage{
data(scen_DJFTas)
}
\description{
scen_DJFTas gives the GCM and RCM which have been used for the 20 climate
projections (obtained with the RCP8.5)
}
\author{
Guillaume Evin \email{guillaume.evin@inrae.fr}
}
\keyword{data}
