% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/description-data.r
\docType{data}
\name{scen_SWE}
\alias{scen_SWE}
\title{data.frame indicating which GCM, RCM and RCP scenarios have been used to produce the 18
projections of SWE maxima.}
\format{
matrix 18 scenarios x 149 years
}
\usage{
data(scen_SWE)
}
\description{
SWE maxima are provided for 18 projections obtained with 9 different combinations of GCMs and RCMs
and two emission scenarios (RCP4.5 and RCP8.5).
}
\references{
Evin, G., E. Le Roux, E. Kamir, and S. Morin. « Estimating changes in 
extreme snow load in Europe as a function of global warming levels ». Cold Regions
 Science and Technology 231 (2025): 104424. https://doi.org/10.1016/j.coldregions.2025.104424.
}
\author{
Guillaume Evin \email{guillaume.evin@inrae.fr}
}
\keyword{data}
