% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/description-data.r
\docType{data}
\name{scenAvail}
\alias{scenAvail}
\title{scenAvail gives the GCM and RCM which have been used for the 20 climate
projections}
\format{
data.frame with 20 rows and two columns: GCM and RCM
}
\usage{
data(scenAvail)
}
\description{
scenAvail gives the GCM and RCM which have been used for the 20 climate
projections
}
\author{
Guillaume Evin \email{guillaume.evin@inrae.fr}
}
\keyword{data}
