% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/description-data.r
\docType{data}
\name{X_SWE_WL}
\alias{X_SWE_WL}
\title{Annual warming levels simulated by different CMIP5 GCMs corresponding to the
18 projections Y_SWE}
\format{
matrix 18 scenarios x 149 years
}
\usage{
data(X_SWE_WL)
}
\description{
Annual warming levels at the planetary scales simulated by different CMIP5
GCMs for the period 1951-2099. Warming levels are obtained with respect to
the year 1860 (common starting year of the CMIP5 simulations). See Evin et al. (2025)
for further details.
}
\references{
Evin, G., E. Le Roux, E. Kamir, and S. Morin. « Estimating changes in 
extreme snow load in Europe as a function of global warming levels ». Cold Regions
 Science and Technology 231 (2025): 104424. https://doi.org/10.1016/j.coldregions.2025.104424.
}
\author{
Guillaume Evin \email{guillaume.evin@inrae.fr}
}
\keyword{data}
