% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/description-data.r
\docType{data}
\name{X_time_mat}
\alias{X_time_mat}
\title{Years 1971-2099 repeated for the 20 scenarios}
\format{
matrix 20 scenarios x 129 years
}
\usage{
data(X_time_mat)
}
\description{
Years 1971-2099 repeated for the 20 scenarios
}
\author{
Guillaume Evin \email{guillaume.evin@inrae.fr}
}
\keyword{data}
