% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/description-data.r
\docType{data}
\name{X_time_vec}
\alias{X_time_vec}
\title{X_time_vec gives the years corr. to Y, i.e. from 1971 to 2099}
\format{
vector of length 129
}
\usage{
data(X_time_vec)
}
\description{
X_time_vec gives the years corr. to Y, i.e. from 1971 to 2099
}
\author{
Guillaume Evin \email{guillaume.evin@inrae.fr}
}
\keyword{data}
