% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/description-data.r
\docType{data}
\name{Xfut_globaltas}
\alias{Xfut_globaltas}
\title{Vector of of future warming levels}
\format{
vector of length 13
}
\usage{
data(Xfut_globaltas)
}
\description{
Equally spaced vector of of future warming levels
}
\author{
Guillaume Evin \email{guillaume.evin@inrae.fr}
}
\keyword{data}
