% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/description-data.r
\docType{data}
\name{Xfut_time}
\alias{Xfut_time}
\title{Xfut_time is a vector of 11 years equally spaced from 1999 to 2099}
\format{
vectors of length 11
}
\usage{
data(Xfut_time)
}
\description{
Xfut_time is a vector of 11 years equally spaced from 1999 to 2099
}
\author{
Guillaume Evin \email{guillaume.evin@inrae.fr}
}
\keyword{data}
