% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/description-data.r
\docType{data}
\name{Y_DJFTas}
\alias{Y_DJFTas}
\title{Mean winter temperature over CEU with 20 GCM/RCM combinations for 1971-2099}
\format{
matrix 20 scenarios x 129 years
}
\usage{
data(Y_DJFTas)
}
\description{
climate projections of mean winter (DJF) temperature over the SREX region CEU
simulated by 20 combinations of CMIP5 GCMs and RCMs for the period 1971-2099
}
\references{
Seneviratne, S. I. et al. Changes in Climate Extremes and their
Impacts on the Natural Physical Environment, in: Managing the Risks of
Extreme Events and Disasters to Advance Climate Change Adaptation: Special
Report of the Intergovernmental Panel on Climate Change, edited by: Field,
C., Barros, V., Stocker, T., and Dahe, Q., Cambridge University Press,
Cambridge, 109-230, https://doi.org/10.1017/CBO9781139177245.006, 2012
}
\author{
Guillaume Evin \email{guillaume.evin@inrae.fr}
}
\keyword{data}
