% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QUALYPSO.r
\name{fit.climate.response}
\alias{fit.climate.response}
\title{fit.climate.response}
\usage{
fit.climate.response(Y, parSmooth = 1, indexReferenceYear = 1,
  typeChangeVariable = "abs")
}
\arguments{
\item{Y}{matrix ofsimulation chains: \code{nS} x \code{nY}}

\item{parSmooth}{smoothing parameter \code{spar} in \link[stats]{smooth.spline}: varies in [0,1]}

\item{indexReferenceYear}{index of the control year (e.g. 1990)}

\item{typeChangeVariable}{type of change variable: "abs" (absolute) or "rel" (relative)}
}
\value{
list with the following fields:
\itemize{
  \item \strong{phiStar}: climate change response
  \item \strong{etaStar}: internal variability
  \item \strong{phi}: raw trend obtained using \link[stats]{smooth.spline}
  \item \strong{climateResponse}: output from \link[stats]{smooth.spline}
  \item \strong{varInterVariability}: scalar, internal variability component of the MME
}
}
\description{
fit trends for each simulation chain of an ensemble of \code{nS} projections. Each projection can be a time series
of \code{nY} time steps (e.g. number of years)
}
\author{
Guillaume Evin
}
