% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QUALYPSO.r
\name{get.Qmat}
\alias{get.Qmat}
\title{get.Qmat}
\usage{
get.Qmat(p)
}
\arguments{
\item{p}{integer}
}
\value{
\item{matrix}{p x p matrix}
}
\description{
Provide matrix Q derived from a matrix Q* of Helmert contrasts: \deqn{Q = Q^* (Q^{*T} Q^*)^{-1/2}}
See Eq. A6 in Evin et al., 2019.
}
\references{
Evin, G., B. Hingray, J. Blanchet, N. Eckert, S. Morin, and D. Verfaillie (2020)
Partitioning Uncertainty Components of an Incomplete Ensemble of Climate Projections Using Data Augmentation.
Journal of Climate. J. Climate, 32, 2423–2440. <doi:10.1175/JCLI-D-18-0606.1>.
}
\author{
Guillaume Evin
}
