% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QUALYPSO.r
\name{get.Qstar.mat}
\alias{get.Qstar.mat}
\title{get.Qstar.mat}
\usage{
get.Qstar.mat(p)
}
\arguments{
\item{p}{integer}
}
\value{
\item{matrix}{p x (p-1) matrix containing Helmert contrasts}
}
\description{
Provide matrix containing Helmert contrasts (see Eq. A7 in Evin et al., 2019).
}
\references{
Evin, G., B. Hingray, J. Blanchet, N. Eckert, S. Morin, and D. Verfaillie. (2019) <doi:10.1175/JCLI-D-18-0606.1>.

Evin, G., B. Hingray, J. Blanchet, N. Eckert, S. Morin, and D. Verfaillie.
Partitioning Uncertainty Components of an Incomplete Ensemble of Climate Projections Using Data Augmentation.
Journal of Climate. J. Climate, 32, 2423–2440. <doi:10.1175/JCLI-D-18-0606.1>.
}
\author{
Guillaume Evin
}
