% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QUALYPSO.r
\name{plotQUALYPSOgrandmean}
\alias{plotQUALYPSOgrandmean}
\title{plotQUALYPSOgrandmean}
\usage{
plotQUALYPSOgrandmean(QUALYPSOOUT, CIlevel = c(0.025, 0.975),
  lim = NULL, col = "black", xlab = "Years", ylab = "Grand mean",
  addLegend = T, ...)
}
\arguments{
\item{QUALYPSOOUT}{output from \code{\link{QUALYPSO}}}

\item{CIlevel}{probabilities for the credible intervals, default is equal to \code{c(0.025,0.975)}}

\item{lim}{y-axis limits (default is NULL)}

\item{col}{color for the overall mean and the credible interval}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{addLegend}{if TRUE, a legend is added}

\item{...}{additional arguments to be passed to \code{\link[graphics]{plot}}}
}
\description{
Plot prediction of grand mean ensemble. By default, we plot the credible interval corresponding to a probability 0.95.
}
\author{
Guillaume Evin
}
