% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QUALYPSO.r
\name{plotQUALYPSOinternalvar}
\alias{plotQUALYPSOinternalvar}
\title{plotQUALYPSOinternalvar}
\usage{
plotQUALYPSOinternalvar(
  QUALYPSOOUT,
  lim = NULL,
  xlab = "X",
  ylab = "eta*",
  ...
)
}
\arguments{
\item{QUALYPSOOUT}{output from \code{\link{QUALYPSO}}}

\item{lim}{y-axis limits (default is NULL)}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{...}{additional arguments to be passed to \code{\link[graphics]{plot}}}
}
\description{
Plot eta* as a function of X as well as eta* +/- 1.645*sqrt(INTERNALVAR),
where INTERNALVAR is a function of Xfut.
}
\author{
Guillaume Evin
}
