% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QUALYPSO.r
\name{plotQUALYPSOTotalVarianceByScenario}
\alias{plotQUALYPSOTotalVarianceByScenario}
\title{plotQUALYPSOTotalVarianceByScenario}
\usage{
plotQUALYPSOTotalVarianceByScenario(QUALYPSOOUT, iEff, nameScenario,
  probCI = 0.9, col = NULL, ylim = NULL, xlab = "Years",
  ylab = "Change variable", addLegend = TRUE, ...)
}
\arguments{
\item{QUALYPSOOUT}{output from \code{\link{QUALYPSO}}}

\item{iEff}{index in \code{scenAvail} corresponding to the scenarios (e.g. RCP scenarios)}

\item{nameScenario}{name of the scenario to be plotted (as provided in \code{scenAvail})}

\item{probCI}{probability for the dredible interval, =0.9 by default}

\item{col}{colors for each source of uncertainty, the first two colors corresponding to internal variability and residual variability, respectively}

\item{ylim}{y-axis limits}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{addLegend}{if TRUE, a legend is added}

\item{...}{additional arguments to be passed to \code{\link[graphics]{plot}}}
}
\description{
Plot fraction of total variance explained by each source of uncertainty.
}
\author{
Guillaume Evin
}
