% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QUALYPSO.r
\name{plotQUALYPSOTotalVarianceDecomposition}
\alias{plotQUALYPSOTotalVarianceDecomposition}
\title{plotQUALYPSOTotalVarianceDecomposition}
\usage{
plotQUALYPSOTotalVarianceDecomposition(
  QUALYPSOOUT,
  vecEff = NULL,
  col = c("orange", "yellow", "cadetblue1", "blue1", "darkgreen", "darkgoldenrod4",
    "darkorchid1"),
  xlab = "",
  ylab = "\% Total Variance",
  addLegend = TRUE,
  ...
)
}
\arguments{
\item{QUALYPSOOUT}{output from \code{\link{QUALYPSO}}}

\item{vecEff}{vector of indices corresponding to the main effects (NULL by default), so that the order of appearance in the plot can be modified}

\item{col}{colors for each source of uncertainty, the first two colors corresponding to internal variability and residual variability, respectively}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{addLegend}{if TRUE, a legend is added}

\item{...}{additional arguments to be passed to \code{\link[graphics]{plot}}}
}
\description{
Plot fraction of total variance explained by each source of uncertainty.
}
\author{
Guillaume Evin
}
