% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QUALYPSO.r
\name{QUALYPSO.check.option}
\alias{QUALYPSO.check.option}
\title{QUALYPSO.check.option}
\usage{
QUALYPSO.check.option(listOption)
}
\arguments{
\item{listOption}{list of options}
}
\value{
List containing the complete set of options.
}
\description{
Check if input options provided in \code{\link{QUALYPSO}} are valid and assigned default values if missing.
}
\author{
Guillaume Evin
}
