% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QUALYPSO.r
\name{QUALYPSO.process.scenario}
\alias{QUALYPSO.process.scenario}
\title{QUALYPSO.process.scenario}
\usage{
QUALYPSO.process.scenario(scenAvail)
}
\arguments{
\item{scenAvail}{matrix of available combinations \code{nS} x \code{nEff}}
}
\value{
list of preprocessed objects (\code{listEff, scenAvail, scenComp, nEff, nTypeEff, nComp, isMissing, nMissing, iMatchScen,
indexEffInCompScen, Qmat})
}
\description{
Process input scenarios.
}
\author{
Guillaume Evin
}
