% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QUALYPSO.r
\name{QUALYPSO.process.scenario}
\alias{QUALYPSO.process.scenario}
\title{QUALYPSO.process.scenario}
\usage{
QUALYPSO.process.scenario(scenAvail, computeEmpEff)
}
\arguments{
\item{scenAvail}{matrix of available combinations, ex \code{cbind(c('GCM1','GCM1','GCM2'),c('RCM1','RCM2','RCM1'))}}

\item{computeEmpEff}{vector of column indices in scenAvail corresponding to effects which are estimated empirically}
}
\value{
list of objects
}
\description{
Process input scenarios
}
\author{
Guillaume Evin
}
