% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QUALYPSO.r
\name{QUALYPSO}
\alias{QUALYPSO}
\title{QUALYPSO}
\usage{
QUALYPSO(
  Y,
  scenAvail,
  X = NULL,
  Xref = NULL,
  Xfut = NULL,
  iFut = NULL,
  listOption = NULL
)
}
\arguments{
\item{Y}{matrix \code{nS} x \code{nY} or array \code{nG} x \code{nS} x \code{nY} of climate projections}

\item{scenAvail}{matrix of available combinations \code{nS} x \code{nEff}. The number of characteristics \code{nEff} corresponds to the
number of main effects which will be included in the ANOVA model.}

\item{X}{(optional) predictors corresponding to the projections, e.g. time or global temperature.
It can be a vector if the predictor is the same for all scenarios (e.g. \code{X=2001:2100} or
a matrix of the same size as Y is these predictors are different for the scenarios. By default,
a vector \code{1:nY} is created.}

\item{Xref}{(optional) reference/control value of the predictor (e.g. the reference period).
X must be included in the range of values of X. For example, if \code{X=1980:2100}, Xref can be any value
between 1980 and 2100. By default, Xref is taken as the minimum value of X.}

\item{Xfut}{(optional) values of the predictor over which the ANOVA will be applied. It must be
a vector of values within the range of values of X. By default, it corresponds to X if X is a vector,
\code{1:nY} if X is \code{NULL} or a vector of 10 values equally spaced between the minimum and
maximum values of X if X is a matrix.}

\item{iFut}{index in \code{1:length(Xfut)} corresponding to a future predictor value . This index is
necessary when \code{Y} is an array \code{nG} x \code{nS} x \code{nY} available for \code{nG} grid points.
Indeed, in this case, we run QUALYPSO only for one future predictor.}

\item{listOption}{(optional) list of options
\itemize{
  \item \strong{spar}: smoothing parameter \code{spar} in \link[stats]{smooth.spline}: typically (but not necessarily) in (0,1]
  \item \strong{lambda}: smoothing parameter \code{lambda} in \link[stats]{smooth.spline}
  \item \strong{typeChangeVariable}: type of change variable: "abs" (absolute, value by default) or "rel" (relative)
  \item \strong{nBurn}: number of burn-in samples (default: 1000). If \code{nBurn} is too small, the convergence of MCMC chains might not be obtained.
  \item \strong{nKeep}: number of kept samples (default: 2000). If \code{nKeep} is too small, MCMC samples might not be represent correctly the posterior
  distributions of inferred parameters.
  \item \strong{nCluster}: number of clusters used for the parallelization (default: 1). When \code{nCluster} is greater than one, parallelization is used to
  apply \code{QUALYPSO} over multiple time steps or grid points simultaneously.
  \item \strong{quantileCompress}: vector of probabilities (in [0,1]) for which we compute the quantiles from the posterior distributions
   \code{quantileCompress = c(0.005,0.025,0.05,0.1,0.25,0.33,0.5,0.66,0.75,0.9,0.95,0.975,0.995)} by default
}}
}
\value{
list with the following fields:
\itemize{
  \item \strong{CLIMATEESPONSE}: list of climate change responses and corresponding internal variability. Contains \code{phiStar} (climate change responses),
  \code{etaStar} (deviation from the climate change responses as a result of internal variability), and \code{phi} (fitted climate responses)
  \item \strong{BAYES}: list of inferred quantities: quantiles from the posterior distributions are provided
  \itemize{
      \item GRANDMEAN: Grand mean
      \item MAINEFFECT: Main effects
      \item RESIDUALVAR_MEAN: mean of the posterior distribution for the residual variability \eqn{\sigma^2}
      \item RESIDUALVAR_QUANT: quantiles of the posterior distribution for the residual variability \eqn{\sigma^2}
      \item INTERNALVAR: Internal variability (constant over time)
      \item EFFECTVAR: Variability related to the main effects (i.e. variability between the different RCMs, GCMs,..)
      \item TOTALVAR: total variability, i.e. the sum of internal variability, residual variability and
      variability related to the main effect
      \item DECOMPVAR: Decomposition of the total variability for each component
      \item CONTRIB_EACH_EFFECT: Contribution of each individual effect to its component (percentage), e.g. what
      is the contribution of GCM1 to the variability related to GCMs
      \item CHANGEBYEFFECT: For each main effect, mean change by type of effect, i.e. mean change by scenario (RCP4.5)
  }
  \item \strong{POINT}: list of point estimates: mean of the posterior distributions
  \itemize{
      \item GRANDMEAN: Grand mean
      \item MAINEFFECT: Main effects
      \item RESIDUALVAR: Variance of residual errors computed from the differences between the climate change responses
      and the additive anova formula (grand mean + main effects)
      \item INTERNALVAR: Internal variability (constant over time)
      \item EFFECTVAR: Variability related to the main effects (i.e. variability between the different RCMs, GCMs,..)
      \item TOTALVAR: total variability, i.e. the sum of internal variability, residual variability and
      variability related to the main effect
      \item DECOMPVAR: Decomposition of the total variability for each component
      \item CONTRIB_EACH_EFFECT: Contribution of each individual effect to its component (percentage), e.g. what
      is the contribution of GCM1 to the variability related to GCMs
      \item CHANGEBYEFFECT: For each main effect, mean change by type of effect, i.e. mean change by scenario (RCP4.5)
      \item RESERR: differences between the climate change responses and the additive anova formula (grand mean + main effects)
  }

  \item \strong{Xmat}: matrix of predictors
  \item \strong{Xref}: reference predictor value
  \item \strong{Xfut}: future predictor values
  \item \strong{paralType}: type of parallelisation (Time or Grid)
  \item \strong{namesEff}: names of the main effects
  \item \strong{Y}: matrix of available combinations given as inputs
  \item \strong{listOption}: list of options used to obtained these results (obtained from \code{\link{QUALYPSO.check.option}})
  \item \strong{listScenarioInput}: list of scenario characteristics (obtained from \code{\link{QUALYPSO.process.scenario}})
}
}
\description{
Partition uncertainty in climate responses using an ANOVA inferred with a Bayesian approach.
}
\examples{
##########################################################################
# SYNTHETIC SCENARIOS
##########################################################################
# create nS=3 fictive climate scenarios with 2 GCMs and 2 RCMs, for a period of nY=20 years
n=20
t=1:n/n

# GCM effects (sums to 0 for each t)
effGCM1 = t*2
effGCM2 = t*-2

# RCM effects (sums to 0 for each t)
effRCM1 = t*1
effRCM2 = t*-1

# These climate scenarios are a sum of effects and a random gaussian noise
scenGCM1RCM1 = effGCM1 + effRCM1 + rnorm(n=n,sd=0.5)
scenGCM1RCM2 = effGCM1 + effRCM2 + rnorm(n=n,sd=0.5)
scenGCM2RCM1 = effGCM2 + effRCM1 + rnorm(n=n,sd=0.5)
Y.synth = rbind(scenGCM1RCM1,scenGCM1RCM2,scenGCM2RCM1)

# Here, scenAvail indicates that the first scenario is obtained with the combination of the
# GCM "GCM1" and RCM "RCM1", the second scenario is obtained with the combination of
# the GCM "GCM1" and RCM "RCM2" and the third scenario is obtained with the combination
# of the GCM "GCM2" and RCM "RCM1".
scenAvail.synth = data.frame(GCM=c('GCM1','GCM1','GCM2'),RCM=c('RCM1','RCM2','RCM1'))

##########################################################################
# RUN QUALYPSO
##########################################################################
# call main QUALYPSO function: two arguments are mandatory:
# - Y: Climate projections for nS scenarions and nY time steps. if Y is a matrix nS x nY, we
# run QUALYPSO nY times, for each time step. If Y is an array nG x nS x nY, for nG grid points,
# we run QUALYPSO nG times, for each grid point, for one time step specified using the argument
# iFut
# - scenAvail: matrix or data.frame of available combinations nS x nEff. The number of
# characteristics nEff corresponds to the number of main effects which will be included in the
# ANOVA model. In the following example, we have nEff=2 main effects corresponding to the GCMs
# and RCMs.

# Many options can be specified in the argument "listOption". Here, we change the default values
# for nBurn and nKeep in order to speed up computation time for this small example. However, it must
# be noticed that convergence and sampling of the posterior distributions often require higher
# values for these two parameters.
listOption = list(nBurn=100,nKeep=100,quantileCompress=c(0.025,0.5,0.975))

# run QUALYPSO
QUALYPSO.synth = QUALYPSO(Y=Y.synth, scenAvail=scenAvail.synth, X=2001:2020, listOption=listOption)

##########################################################################
# SOME PLOTS
##########################################################################
# plot grand mean
plotQUALYPSOgrandmean(QUALYPSO.synth,xlab="Years")

# plot main GCM effects
plotQUALYPSOeffect(QUALYPSO.synth,iEff=1,xlab="Years")

# plot main RCM effects
plotQUALYPSOeffect(QUALYPSO.synth,iEff=2,xlab="Years")

# plot fraction of total variance for the differences sources of uncertainty
plotQUALYPSOTotalVarianceDecomposition(QUALYPSO.synth,xlab="Years")

# plot mean prediction and total variance with the differences sources of uncertainty
plotQUALYPSOMeanChangeAndUncertainties(QUALYPSO.synth,xlab="Years")

#____________________________________________________________
# EXAMPLE OF QUALYPSO WHEN THE PREDICTOR IS TIME
#____________________________________________________________

# list of options
listOption = list(typeChangeVariable='abs')

# call QUALYPSO
QUALYPSO.time = QUALYPSO(Y=Y,scenAvail=scenAvail,X=X_time_vec,Xref=1999,
                         Xfut=Xfut_time,listOption=listOption)

# grand mean effect
plotQUALYPSOgrandmean(QUALYPSO.time,xlab="Years")

# main GCM effects
plotQUALYPSOeffect(QUALYPSO.time,iEff = 1,xlab="Years")

# main RCM effects
plotQUALYPSOeffect(QUALYPSO.time,iEff = 2,xlab="Years")

# mean change and associated uncertainties
plotQUALYPSOMeanChangeAndUncertainties(QUALYPSO.time,xlab="Years")

# variance decomposition
plotQUALYPSOTotalVarianceDecomposition(QUALYPSO.time,xlab="Years")

#____________________________________________________________
# EXAMPLE OF QUALYPSO WHEN THE PREDICTOR IS THE GLOBAL TEMPERATURE
#____________________________________________________________

# list of options
listOption = list(typeChangeVariable='abs')

# call QUALYPSO
QUALYPSO.globaltas = QUALYPSO(Y=Y,scenAvail=scenAvail,X=X_globaltas,Xref=13,
                              Xfut=Xfut_globaltas,listOption=listOption)

# grand mean effect
plotQUALYPSOgrandmean(QUALYPSO.globaltas,xlab="Global temperature (Celsius)")

# main GCM effects
plotQUALYPSOeffect(QUALYPSO.globaltas,iEff = 1,xlab="Global temperature (Celsius)")

# main RCM effects
plotQUALYPSOeffect(QUALYPSO.globaltas,iEff = 2,xlab="Global temperature (Celsius)")

# mean change and associated uncertainties
plotQUALYPSOMeanChangeAndUncertainties(QUALYPSO.globaltas,xlab="Global temperature (Celsius)")

# variance decomposition
plotQUALYPSOTotalVarianceDecomposition(QUALYPSO.globaltas,xlab="Global temperature (Celsius)")

}
\references{
Evin, G., B. Hingray, J. Blanchet, N. Eckert, S. Morin, and D. Verfaillie.
Partitioning Uncertainty Components of an Incomplete Ensemble of Climate Projections Using Data Augmentation.
Journal of Climate. <doi:10.1175/JCLI-D-18-0606.1>.
}
\author{
Guillaume Evin
}
