% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/description-data.r
\docType{data}
\name{Xfut_globaltas}
\alias{Xfut_globaltas}
\title{Equally spaced vector of simulated global temperatures over the period
1971-2099 for the RCP8.5}
\format{
vector of length 13
}
\usage{
data(Xfut_globaltas)
}
\description{
Equally spaced vector of simulated global temperatures over the period
1971-2099 for the RCP8.5
}
\author{
Guillaume Evin \email{guillaume.evin@inrae.fr}
}
\keyword{data}
